+++ src/frontend/js/components/DashboardModerators/ButtonChangeRoleCta/index.tsx
@@ -1,0 +1,48 @@
+import React from 'react';
+import { FormattedMessage } from 'react-intl';
+import { messagesDashboardModerators } from '../messages';
+import { Actions, Role } from '../../../types/Enums';
+import { User } from '../../../types/User';
+import { updateUser } from '../../../data/fetchApi';
+
+export interface ButtonChangeRoleCtaProps {
+  user: User;
+  action: Actions;
+  onChange: () => void;
+}
+
+export const ButtonChangeRoleCta = ({
+  user,
+  action,
+  onChange,
+}: ButtonChangeRoleCtaProps) => {
+  const message =
+    action === Actions.PROMOTE
+      ? messagesDashboardModerators.promoteModerator
+      : messagesDashboardModerators.revokeModerator;
+
+  const updateStudent = async () => {
+    // Depending on action submit the requested role
+    switch (action) {
+      case Actions.PROMOTE:
+        user.role = Role.MODERATOR;
+        break;
+      case Actions.REVOKE:
+        user.role = Role.STUDENT;
+        break;
+    }
+    const content = await updateUser(user);
+    onChange();
+    return content;
+  };
+
+  return (
+    <button
+      onClick={updateStudent}
+      id="add-moderator__cta"
+      className="btn btn-primary"
+    >
+      <FormattedMessage {...message} />
+    </button>
+  );
+};