+++ src/frontend/js/types/Enums.ts
@@ -1,0 +1,9 @@
+export enum Actions {
+  REVOKE = 'revoke',
+  PROMOTE = 'promote',
+}
+
+export enum Role {
+  STUDENT = 'student',
+  MODERATOR = 'moderator',
+}